+++ src/main/java/net/mcreator/allaboutengie/procedures/EpicSharkoAprilFoolsTameRCProcedure.java
@@ -31,7 +31,7 @@ public class EpicSharkoAprilFoolsTameRCProcedure {
 					_entity.setHealth(entity instanceof LivingEntity _livEnt ? _livEnt.getMaxHealth() : -1);
 			} else {
 				if (entity instanceof LivingEntity _entity)
-					_entity.setHealth((float) ((entity instanceof LivingEntity _livEnt ? _livEnt.getHealth() : -1) + 5));
+					_entity.setHealth((entity instanceof LivingEntity _livEnt ? _livEnt.getHealth() : -1) + 5);
 				if (world instanceof ServerLevel _level)
 					_level.sendParticles(ParticleTypes.HEART, x, y, z, 5, 1, 1, 1, 1);
 			}
